refactor && clean
