+++ vm2/chain/Cargo.toml
@@ -17,6 +17,6 @@ starcoin-vm2-statedb = { workspace = true }
 starcoin-vm2-types = { workspace = true }
 starcoin-vm2-vm-types = { workspace = true }
 starcoin-vm2-state-api = { workspace = true }
-starcoin-exec-merge = { workspace = true, features = ["statedb"] }
+starcoin-exec-merge = { workspace = true }
 starcoin-vm2-crypto = { workspace = true }
 starcoin_vm_runtime = { path = "../vm-runtime", package = "starcoin-vm-runtime" }
